fix some small issues

+++ tasks/1429/java/test/TaskTest.java
@@ -89,7 +89,7 @@ public class TaskTest {
       Task t = new Task();
       Task.Circle c1 = t.new Circle(0, 0, 3);
       Task.Circle c2 = t.new Circle(4, 0, 2);
-      LinkedList<Task.Pair<Float, Float>> p = c1.calculateVertex(c2);
+      LinkedList<Task.Pair<Double, Double>> p = c1.calculateVertex(c2);
       assertEquals(2, p.size());
       
       c1 = t.new Circle(-5, -5, 3);
@@ -101,9 +101,9 @@ public class TaskTest {
       c2 = t.new Circle(4, 3, 2);
       p = c1.calculateVertex(c2);
       assertEquals(1, p.size());
-      Task.Pair<Float, Float> p1 = p.get(0);
+      Task.Pair<Double, Double> p1 = p.get(0);
       p=c2.calculateVertex(c1);
-      Task.Pair<Float, Float> p2 = p.get(0);
+      Task.Pair<Double, Double> p2 = p.get(0);
       assertEquals(p1, p2);
       
 //      c1 = t.new Circle(0, 0, 3);
